Re #17061. Removed bad cast from int64 to int32, fixing OSX issues?
Previously the index type was size_t, i.e., also 64 bit, but actually
contained the spectrum number, which was 32 bit, so casting to an int
did not cause trouble. Now we are using the this pointer as index, and
the values exceed 32 bit. This might explain the failing unit tests on
OSX (clang).

+++ Framework/Kernel/inc/MantidKernel/MRUList.h
@@ -138,7 +138,7 @@ public:
   void deleteIndex(const uintptr_t index) {
     std::lock_guard<std::mutex> _lock(m_mutex);
 
-    auto it = il.template get<1>().find((int)index);
+    auto it = il.template get<1>().find(index);
     if (it != il.template get<1>().end()) {
       delete (*it);
       il.template get<1>().erase(it);
@@ -157,7 +157,7 @@ public:
   T *find(const uintptr_t index) const {
     std::lock_guard<std::mutex> _lock(m_mutex);
 
-    auto it = il.template get<1>().find(int(index));
+    auto it = il.template get<1>().find(index);
     if (it == il.template get<1>().end()) {
       return nullptr;
     } else {
